server/requirements/dev.txt
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Django==2.2.4
django-filter==2.2.0
djangorestframework==3.10.2
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
Markdown==3.1.1
mccabe==0.6.1
mtnmomo==3.0.1
pep8==1.7.1
phonenumbers==8.10.17
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pylint==2.3.1
python-dotenv==0.10.3
pytz==2019.2
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.24.3
wrapt==1.11.2
